% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWordCloud.R
\name{createTermDocumentMatrix}
\alias{createTermDocumentMatrix}
\title{Create Term-Document Matrix}
\usage{
createTermDocumentMatrix(x, type = c("text", "url", "file"),
  lang = "english", excludeWords = NULL, textStemming = FALSE)
}
\arguments{
\item{x:}{character string (plain text, web url, txt file path)}

\item{type:}{specify whether x is a plain text, a web page url or a file path}

\item{lang:}{the language of the text}

\item{excludeWords:}{a vector of words to exclude from the text}

\item{textStemming:}{reduces words to their root form}
}
\value{
a list(tdm, freqTable)
}
\description{
Function to create a term-document matrix based on a text file/object.
}
\references{
\url{http://www.sthda.com/english/wiki/word-cloud-generator-in-r-one-killer-function-to-do-everything-you-need}
}
\author{
Gareth Burns
}
