% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{html_to_text}
\alias{html_to_text}
\title{Download and parse webpage}
\usage{
html_to_text(url)
}
\arguments{
\item{url}{url of website}
}
\value{
character vestor of text from website
}
\description{
Function to create a extraction text from website
}
\references{
\url{http://www.sthda.com/english/wiki/word-cloud-generator-in-r-one-killer-function-to-do-everything-you-need}
}
\author{
Gareth Burns
}
